Completed updater

+++ putosSA/computer/home_system/update.lua
@@ -1,3 +1,14 @@
-fs.delete("home.lua")
-shell.run("wget", "download", "https://raw.githubusercontent.com/Kolterdyx/computercraft/master/putosSA/computer/9/home.lua")
+if fs.exists("home.lua") then
+    fs.delete("home.lua")
+end
+shell.run("wget", "https://raw.githubusercontent.com/Kolterdyx/computercraft/master/putosSA/computer/home_system/home.lua")
+if not fs.exists("table_saver.lua") then
+    shell.run("wget", "https://raw.githubusercontent.com/Kolterdyx/computercraft/master/putosSA/computer/home_system/table_saver.lua")
+end
+if fs.exists("startup.lua") then
+    fs.delete("startup.lua")
+end
+shell.run("wget", "https://raw.githubusercontent.com/Kolterdyx/computercraft/master/putosSA/computer/home_system/startup.lua")
+
+
 os.reboot()
